feat(components): integrate RenderBuffer to skip redundant renders
Introduces `RenderBuffer` support across all Preswald components to prevent unnecessary
re-renders and frontend updates when component state has not changed. Components now
use `generate_stable_id()` to ensure deterministic ids based on callsite or user-supplied
identifiers, enabling consistent diffing between reruns.

Additionally, common service logic has been extracted into a shared `BasePreswaldService`, now inherited by both
`ServerPreswaldService` and `VirtualPreswaldService`.

- All components now use `should_render()` to conditionally append
- Debug logs are now guarded to avoid expensive serialization
- `RenderBuffer` hashes values internally and accepts either raw objects or precomputed hashes
