+++ preswald/interfaces/components.py
@@ -1,4 +1,5 @@
 # Standard Library
+import asyncio
 import base64
 import hashlib
 import io
@@ -9,6 +10,7 @@ import uuid
 from typing import Dict, List, Optional
 
 # Third-Party
+from inspect import currentframe, getframeinfo
 import matplotlib.pyplot as plt
 import numpy as np
 import pandas as pd
@@ -39,19 +41,26 @@ def alert(message: str, level: str = "info", size: float = 1.0) -> str:
     """Create an alert component."""
     service = PreswaldService.get_instance()
 
-    id = generate_id("alert")
-    logger.debug(f"Creating alert component with id {id}, message: {message}")
+    component_id = generate_stable_id("alert")
+    logger.debug(f"Creating alert component with id {component_id}, message: {message}")
+
     component = {
         "type": "alert",
-        "id": id,
+        "id": component_id,
         "message": message,
         "level": level,
         "size": size,
     }
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
-    return message
 
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
+    return message
 
 def button(
     label: str,
@@ -62,7 +71,7 @@ def button(
 ) -> bool:
     """Create a button component that returns True when clicked."""
     service = PreswaldService.get_instance()
-    component_id = generate_id_by_label("button", label)
+    component_id = generate_stable_id("button")
 
     # Get current state or use default
     current_value = service.get_component_state(component_id)
@@ -81,7 +90,14 @@ def button(
         "onClick": True,  # Always enable click handling
     }
 
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return current_value
 
 
@@ -90,7 +106,7 @@ def chat(source: str, table: Optional[str] = None) -> Dict:
     service = PreswaldService.get_instance()
 
     # Create a consistent ID based on the source
-    component_id = f"chat-{hashlib.md5(str(source).encode()).hexdigest()[:8]}"
+    component_id = generate_stable_id("chat")
 
     # Get current state or initialize empty
     current_state = service.get_component_state(component_id)
@@ -135,8 +151,13 @@ def chat(source: str, table: Optional[str] = None) -> Dict:
         },
     }
 
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
     return component
 
 
@@ -145,7 +166,7 @@ def checkbox(label: str, default: bool = False, size: float = 1.0) -> bool:
     service = PreswaldService.get_instance()
 
     # Create a consistent ID based on the label
-    component_id = generate_id_by_label("checkbox", label)
+    component_id = generate_stable_id("checkbox")
 
     # Get current state or use default
     current_value = service.get_component_state(component_id)
@@ -160,8 +181,15 @@ def checkbox(label: str, default: bool = False, size: float = 1.0) -> bool:
         "value": current_value,
         "size": size,
     }
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return current_value
 
 
@@ -243,11 +271,18 @@ def checkbox(label: str, default: bool = False, size: float = 1.0) -> bool:
 def image(src, alt="Image", size=1.0):
     """Create an image component."""
     service = PreswaldService.get_instance()
-    id = generate_id("image")
-    logger.debug(f"Creating image component with id {id}, src: {src}")
-    component = {"type": "image", "id": id, "src": src, "alt": alt, "size": size}
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+    component_id = generate_stable_id("image")
+    logger.debug(f"Creating image component with id {component_id}, src: {src}")
+    component = {"type": "image", "id": component_id, "src": src, "alt": alt, "size": size}
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return component
 
 
@@ -266,7 +301,7 @@ def matplotlib(fig: Optional[plt.Figure] = None, label: str = "plot") -> str:
     img_b64 = base64.b64encode(buf.read()).decode()
 
     # Generate a unique component ID based on the label
-    component_id = f"matplotlib-{hashlib.md5(label.encode()).hexdigest()[:8]}"
+    component_id = generate_stable_id("matplotlib")
 
     component = {
         "type": "matplotlib",
@@ -275,7 +310,13 @@ def matplotlib(fig: Optional[plt.Figure] = None, label: str = "plot") -> str:
         "image": img_b64,  # Store the image data
     }
 
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
 
     return component_id  # Returning ID for potential tracking
 
@@ -300,7 +341,7 @@ def playground(
     service = PreswaldService.get_instance()
 
     # Generate a unique component ID using the label's hash
-    component_id = f"playground-{hashlib.md5(label.encode()).hexdigest()[:8]}"
+    component_id = generate_stable_id("playground")
 
     logger.debug(
         f"Creating playground component with id {component_id}, label: {label}"
@@ -373,8 +414,13 @@ def playground(
         "data": {"columnDefs": column_defs, "rowData": processed_data or []},
     }
 
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
 
     # Return the raw DataFrame
     return data
@@ -388,14 +434,14 @@ def plotly(fig, size: float = 1.0) -> Dict:  # noqa: C901
         fig: A Plotly figure object.
     """
     service = PreswaldService.get_instance()
+    component_id = generate_stable_id("plot")
+
     try:
         import time
 
         start_time = time.time()
         logger.debug("[PLOTLY] Starting plotly render")
-
-        id = generate_id("plot")
-        logger.debug(f"[PLOTLY] Created plot component with id {id}")
+        logger.debug(f"[PLOTLY] Created plot component with id {component_id}")
 
         # Optimize the figure for web rendering
         optimize_start = time.time()
@@ -485,7 +531,7 @@ def plotly(fig, size: float = 1.0) -> Dict:  # noqa: C901
 
         component = {
             "type": "plot",
-            "id": id,
+            "id": component_id,
             "data": {
                 "data": serializable_fig_dict.get("data", []),
                 "layout": serializable_fig_dict.get("layout", {}),
@@ -506,18 +552,25 @@ def plotly(fig, size: float = 1.0) -> Dict:  # noqa: C901
         json.dumps(component)
         logger.debug(f"[PLOTLY] JSON verification took {time.time() - json_start:.3f}s")
 
-        logger.debug(f"[PLOTLY] Plot data created successfully for id {id}")
+        logger.debug(f"[PLOTLY] Plot data created successfully for id {component_id}")
         logger.debug(
             f"[PLOTLY] Total plotly render took {time.time() - start_time:.3f}s"
         )
-        service.append_component(component)
+
+        if service.should_render(component_id, component):
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"Created component: {component}")
+            service.append_component(component)
+        else:
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"No changes detected. skipping append for component {component}")
         return component
 
     except Exception as e:
         logger.error(f"[PLOTLY] Error creating plot: {e!s}", exc_info=True)
         error_component = {
             "type": "plot",
-            "id": id,
+            "id": component_id,
             "error": f"Failed to create plot: {e!s}",
         }
         service.append_component(error_component)
@@ -528,17 +581,24 @@ def progress(label: str, value: float = 0.0, size: float = 1.0) -> float:
     """Create a progress component."""
     service = PreswaldService.get_instance()
 
-    id = generate_id("progress")
-    logger.debug(f"Creating progress component with id {id}, label: {label}")
+    component_id = generate_stable_id("progress")
+    logger.debug(f"Creating progress component with id {component_id}, label: {label}")
     component = {
         "type": "progress",
-        "id": id,
+        "id": component_id,
         "label": label,
         "value": value,
         "size": size,
     }
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return value
 
 
@@ -548,7 +608,7 @@ def selectbox(
     """Create a select component with consistent ID based on label."""
     service = PreswaldService.get_instance()
 
-    component_id = generate_id_by_label("selectbox", label)
+    component_id = generate_stable_id("selectbox")
     current_value = service.get_component_state(component_id)
     if current_value is None:
         current_value = (
@@ -563,15 +623,32 @@ def selectbox(
         "value": current_value,
         "size": size,
     }
-    service.append_component(component)
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return current_value
 
 
 def separator() -> Dict:
     """Create a separator component that forces a new row."""
     service = PreswaldService.get_instance()
-    component = {"type": "separator", "id": str(uuid.uuid4())}
-    service.append_component(component)
+    component_id = generate_stable_id("separator")
+    component = {"type": "separator", "id": component_id}
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return component
 
 
@@ -587,7 +664,7 @@ def slider(
     service = PreswaldService.get_instance()
 
     # Create a consistent ID based on the label
-    component_id = generate_id_by_label("slider", label)
+    component_id = generate_stable_id("slider")
 
     # Get current state or use default
     current_value = service.get_component_state(component_id)
@@ -605,7 +682,14 @@ def slider(
         "size": size,
     }
 
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return current_value
 
 
@@ -624,7 +708,7 @@ def spinner(
         size: Component width (1.0 = full width)
     """
     service = PreswaldService.get_instance()
-    component_id = generate_id("spinner")
+    component_id = generate_stable_id("spinner")
 
     component = {
         "type": "spinner",
@@ -635,18 +719,30 @@ def spinner(
         "size": size,
     }
 
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return None
 
 
 def sidebar(defaultopen: bool = False):
     """Create a sidebar component."""
     service = PreswaldService.get_instance()
-    id = generate_id("sidebar")
-    logger.debug(f"Creating sidebar component with id {id}")
-    component = {"type": "sidebar", "id": id, "defaultopen": defaultopen}
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+    component_id = generate_stable_id("sidebar")
+    logger.debug(f"Creating sidebar component with id {component_id}")
+    component = {"type": "sidebar", "id": component_id, "defaultopen": defaultopen}
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
     return component
 
 
@@ -663,8 +759,8 @@ def table(
     Returns:
         Dict: Component metadata and processed data.
     """
-    id = generate_id("table")
-    logger.debug(f"Creating table component with id {id}")
+    component_id = generate_stable_id("table")
+    logger.debug(f"Creating table component with id {component_id}")
     service = PreswaldService.get_instance()
 
     try:
@@ -712,7 +808,7 @@ def table(
         # Create AG Grid compatible component structure
         component = {
             "type": "table",
-            "id": id,
+            "id": component_id,
             "props": {
                 "columnDefs": column_defs,
                 "rowData": processed_data,
@@ -720,17 +816,21 @@ def table(
             },
         }
 
-        # Verify JSON serialization before returning
-        json.dumps(component)
-        logger.debug(f"Created AG Grid table component: {component}")
-        service.append_component(component)
+        if service.should_render(component_id, component):
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"Created component: {component}")
+            service.append_component(component)
+        else:
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"No changes detected. skipping append for component {component}")
+
         return component
 
     except Exception as e:
         logger.error(f"Error creating table component: {e!s}")
         error_component = {
             "type": "table",
-            "id": id,
+            "id": component_id,
             "props": {
                 "columnDefs": [],
                 "rowData": [],
@@ -740,21 +840,25 @@ def table(
         service.append_component(error_component)
         return error_component
 
-
 def text(markdown_str: str, size: float = 1.0) -> str:
     """Create a text/markdown component."""
     service = PreswaldService.get_instance()
-    id = generate_id("text")
-    logger.debug(f"Creating text component with id {id}")
+    component_id = generate_stable_id("text")
     component = {
         "type": "text",
-        "id": id,
+        "id": component_id,
         "markdown": markdown_str,
         "value": markdown_str,
         "size": size,
     }
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
     return markdown_str
 
 
@@ -776,8 +880,8 @@ def text_input(
         str: Current value of the input
     """
     service = PreswaldService.get_instance()
-    component_id = generate_id_by_label("text_input", label)
-
+    component_id = generate_stable_id("text_input")
+    logger.debug(f"Created component with id: {component_id}")
     # Get current state or use default
     current_value = service.get_component_state(component_id)
     if current_value is None:
@@ -792,18 +896,32 @@ def text_input(
         "size": size,
     }
 
-    service.append_component(component)
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return current_value
 
 
 def topbar() -> Dict:
     """Creates a topbar component."""
     service = PreswaldService.get_instance()
-    id = generate_id("topbar")
-    logger.debug(f"Creating topbar component with id {id}")
-    component = {"type": "topbar", "id": id}
-    logger.debug(f"Created component: {component}")
-    service.append_component(component)
+    component_id = generate_stable_id("topbar")
+    logger.debug(f"Creating topbar component with id {component_id}")
+    component = {"type": "topbar", "id": component_id}
+
+    if service.should_render(component_id, component):
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"Created component: {component}")
+        service.append_component(component)
+    else:
+        if logger.isEnabledFor(logging.DEBUG):
+            logger.debug(f"No changes detected. skipping append for component {component}")
+
     return component
 
 
@@ -816,6 +934,7 @@ def workflow_dag(workflow: Workflow, title: str = "Workflow Dependency Graph") -
         title: Optional title for the visualization
     """
     service = PreswaldService.get_instance()
+    component_id = generate_stable_id("dag")
     try:
         from .workflow import WorkflowAnalyzer
 
@@ -838,10 +957,9 @@ def workflow_dag(workflow: Workflow, title: str = "Workflow Dependency Graph") -
             )
 
         # Create the component with the correct type and data structure
-        id = generate_id("dag")
         component = {
             "type": "dag",  # Changed from "plot" to "dag"
-            "id": id,
+            "id": component_id,
             "data": {
                 "data": [
                     {
@@ -854,8 +972,15 @@ def workflow_dag(workflow: Workflow, title: str = "Workflow Dependency Graph") -
             },
         }
 
-        logger.debug(f"[WORKFLOW_DAG] Created DAG component with id {id}")
-        service.append_component(component)
+        logger.debug(f"[WORKFLOW_DAG] Created DAG component with id {component_id}")
+        if service.should_render(component_id, component):
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"Created component: {component}")
+            service.append_component(component)
+        else:
+            if logger.isEnabledFor(logging.DEBUG):
+                logger.debug(f"No changes detected. skipping append for component {component}")
+
         return component
 
     except Exception as e:
@@ -864,7 +989,7 @@ def workflow_dag(workflow: Workflow, title: str = "Workflow Dependency Graph") -
         )
         error_component = {
             "type": "dag",  # Changed from "plot" to "dag"
-            "id": generate_id("dag"),
+            "id": component_id,
             "error": f"Failed to create DAG visualization: {e!s}",
         }
         service.append_component(error_component)
@@ -896,116 +1021,129 @@ def convert_to_serializable(obj):
         return obj.item()
     return obj
 
-
-def generate_id(prefix: str = "component") -> str:
-    """Generate a unique ID for a component."""
-    return f"{prefix}-{uuid.uuid4().hex[:8]}"
-
-
-def generate_id_by_label(prefix: str, label: str) -> str:
+def generate_stable_id(prefix: str = "component", identifier: Optional[str] = None) -> str:
     """
-    Generate a deterministic component ID based on a label string.
+    Generate a stable and deterministic component ID based on either a user-provided identifier 
+    or the source code callsite.
 
-    Useful for components like Fastplotlib or Slider where the same label should result
-    in the same component ID across rerenders.
+    This is useful for ensuring that components retain the same ID across script reruns, enabling 
+    reliable caching, state restoration, and render diffing.
 
     Args:
-        prefix (str): The component type prefix, e.g. 'slider' or 'fastplotlib'.
-        label (str): The label to hash and include in the ID.
+        prefix (str): A prefix to distinguish the type of component (e.g., "text", "plot").
+        identifier (Optional[str]): An optional string used to generate a stable hash. This can 
+                                    be a label, index, or any meaningful identifier. If not provided, 
+                                    the callsite (filename and line number) will be used.
 
     Returns:
-        str: A stable ID like 'slider-ab12cd34'.
+        str: A stable component ID in the format "<prefix>-<hash>", where <hash> is the first 
+             8 characters of an MD5 hash derived from the identifier or callsite.
+
+    Notes:
+        - If no identifier is provided, the function inspects the call stack and hashes the 
+          file name and line number where `generate_stable_id` was originally invoked.
+        - This makes it easy to write deterministic scripts without manually assigning IDs, 
+          while still supporting manual overrides.
     """
-    if not label:
-        return generate_id(prefix)
-    hashed = hashlib.md5(label.lower().encode()).hexdigest()[:8]
+    def get_callsite_id():
+        frame = currentframe()
+        for _ in range(3):  # skip generate_stable_id, text(), etc.
+            if frame is not None:
+                frame = frame.f_back
+        info = getframeinfo(frame)
+        return f"{info.filename}:{info.lineno}"
+
+    if identifier:
+        hashed = hashlib.md5(identifier.lower().encode()).hexdigest()[:8]
+    else:
+        identifier = get_callsite_id()
+        hashed = hashlib.md5(identifier.encode()).hexdigest()[:8]
     return f"{prefix}-{hashed}"
 
+async def render_and_send_fastplotlib(
+    fig: "fplt.Figure",
+    component_id: str,
+    label: str,
+    size: float,
+    client_id: str,
+    data_hash: str,
+) -> Optional[str]:
+    """
+    Asynchronously renders a Fastplotlib figure to an offscreen canvas, encodes it as a PNG,
+    and streams the resulting image data via WebSocket to the connected frontend client.
 
-# async def render_and_send_fastplotlib(
-#     fig: "fplt.Figure",
-#     component_id: str,
-#     label: str,
-#     size: float,
-#     client_id: str,
-#     data_hash: str,
-# ) -> Optional[str]:
-#     """
-#     Asynchronously renders a Fastplotlib figure to an offscreen canvas, encodes it as a PNG,
-#     and streams the resulting image data via WebSocket to the connected frontend client.
-#
-#     This helper function handles rendering logic, alpha-blending, and ensures robust error
-#     handling. It updates the component state after successfully sending the image data.
-#
-#     Args:
-#         fig (fplt.Figure): The fully configured Fastplotlib figure instance to render.
-#         component_id (str): Unique identifier for the component instance receiving this image.
-#         label (str): Human-readable label describing the component (for logging/debugging).
-#         size (float): Relative size of the component in the UI layout (0.0-1.0).
-#         client_id (str): WebSocket client identifier to route the rendered image correctly.
-#         data_hash (str): SHA-256 hash representing the figure state, used for cache invalidation.
-#
-#     Returns:
-#         str: Returns "Render failed" if framebuffer blending fails, otherwise None.
-#
-#     Raises:
-#         Logs and handles any exceptions internally without raising further.
-#     """
-#     service = PreswaldService.get_instance()
-#
-#     fig.show()  # must call even in offscreen mode to initialize GPU resources
-#
-#     # manually render the scene for all subplots
-#     for subplot in fig:
-#         subplot.viewport.render(subplot.scene, subplot.camera)
-#
-#     # read from the framebuffer
-#     try:
-#         fig.canvas.request_draw()
-#         raw_img = np.asarray(fig.renderer.target.draw())
-#
-#         if raw_img.ndim != 3 or raw_img.shape[2] != 4:
-#             raise ValueError(f"Unexpected image shape: {raw_img.shape}")
-#
-#         # handle alpha blending
-#         alpha = raw_img[..., 3:4] / 255.0
-#         rgb = (raw_img[..., :3] * alpha + (1 - alpha) * 255).astype(np.uint8)
-#
-#     except Exception as e:
-#         logger.error(f"Framebuffer blending failed for {component_id}: {e}")
-#         return "Render failed"
-#
-#     # encode image to PNG
-#     img_buf = io.BytesIO()
-#     Image.fromarray(rgb).save(img_buf, format="PNG")
-#     png_bytes = img_buf.getvalue()
-#
-#     # handle websocket communication
-#     client_websocket = service.websocket_connections.get(client_id)
-#     if client_websocket:
-#         packed_msg = msgpack.packb(
-#             {
-#                 "type": "image_update",
-#                 "component_id": component_id,
-#                 "format": "png",
-#                 "label": label,
-#                 "size": size,
-#                 "data": png_bytes,
-#             },
-#             use_bin_type=True,
-#         )
-#
-#         try:
-#             await client_websocket.send_bytes(packed_msg)
-#             await service.handle_client_message(
-#                 client_id,
-#                 {
-#                     "type": "component_update",
-#                     "states": {f"{component_id}_img_hash": data_hash},
-#                 },
-#             )
-#             logger.debug(f"✅ Sent {component_id} image to client {client_id}")
-#         except Exception as e:
-#             logger.error(f"WebSocket send failed for {component_id}: {e}")
-#     else:
-#         logger.warning(f"No active WebSocket found for client ID: {client_id}")
+    This helper function handles rendering logic, alpha-blending, and ensures robust error
+    handling. It updates the component state after successfully sending the image data.
+
+    Args:
+        fig (fplt.Figure): The fully configured Fastplotlib figure instance to render.
+        component_id (str): Unique identifier for the component instance receiving this image.
+        label (str): Human-readable label describing the component (for logging/debugging).
+        size (float): Relative size of the component in the UI layout (0.0-1.0).
+        client_id (str): WebSocket client identifier to route the rendered image correctly.
+        data_hash (str): SHA-256 hash representing the figure state, used for cache invalidation.
+
+    Returns:
+        str: Returns "Render failed" if framebuffer blending fails, otherwise None.
+
+    Raises:
+        Logs and handles any exceptions internally without raising further.
+    """
+    service = PreswaldService.get_instance()
+
+    fig.show()  # must call even in offscreen mode to initialize GPU resources
+
+    # manually render the scene for all subplots
+    for subplot in fig:
+        subplot.viewport.render(subplot.scene, subplot.camera)
+
+    # read from the framebuffer
+    try:
+        fig.canvas.request_draw()
+        raw_img = np.asarray(fig.renderer.target.draw())
+
+        if raw_img.ndim != 3 or raw_img.shape[2] != 4:
+            raise ValueError(f"Unexpected image shape: {raw_img.shape}")
+
+        # handle alpha blending
+        alpha = raw_img[..., 3:4] / 255.0
+        rgb = (raw_img[..., :3] * alpha + (1 - alpha) * 255).astype(np.uint8)
+
+    except Exception as e:
+        logger.error(f"Framebuffer blending failed for {component_id}: {e}")
+        return "Render failed"
+
+    # encode image to PNG
+    img_buf = io.BytesIO()
+    Image.fromarray(rgb).save(img_buf, format="PNG")
+    png_bytes = img_buf.getvalue()
+
+    # handle websocket communication
+    client_websocket = service.websocket_connections.get(client_id)
+    if client_websocket:
+        packed_msg = msgpack.packb(
+            {
+                "type": "image_update",
+                "component_id": component_id,
+                "format": "png",
+                "label": label,
+                "size": size,
+                "data": png_bytes,
+            },
+            use_bin_type=True,
+        )
+
+        try:
+            await client_websocket.send_bytes(packed_msg)
+            await service.handle_client_message(
+                client_id,
+                {
+                    "type": "component_update",
+                    "states": {f"{component_id}_img_hash": data_hash},
+                },
+            )
+            logger.debug(f"✅ Sent {component_id} image to client {client_id}")
+        except Exception as e:
+            logger.error(f"WebSocket send failed for {component_id}: {e}")
+    else:
+        logger.warning(f"No active WebSocket found for client ID: {client_id}")
